refactor: extract starter visible pilot commands
