+++ examples/cli-starter-visible-driver-family-command-modularization-smoke.py
@@ -9,7 +9,9 @@ from pathlib import Path
 
 ROOT = Path(__file__).resolve().parents[1]
 STARTER = ROOT / "loopx" / "cli_commands" / "starter.py"
+VISIBLE_COMMON = ROOT / "loopx" / "cli_commands" / "starter_visible_common.py"
 VISIBLE_DRIVER = ROOT / "loopx" / "cli_commands" / "starter_visible_driver.py"
+VISIBLE_PILOT = ROOT / "loopx" / "cli_commands" / "starter_visible_pilot.py"
 RUNTIME_IDLE = ROOT / "loopx" / "cli_commands" / "starter_runtime_idle.py"
 INIT = ROOT / "loopx" / "cli_commands" / "__init__.py"
 VISIBLE_HELP_FIXTURE = (
@@ -20,12 +22,14 @@ VISIBLE_PROOF_FIXTURE = (
 )
 
 
-VISIBLE_DRIVER_COMMANDS = {
+VISIBLE_PILOT_COMMANDS = {
     "codex-cli-one-message-loop-pilot": ["--proof-fixture", "--allow-headless-fallback"],
     "codex-cli-visible-local-driver-pilot": ["--idle-fixture", "--allow-headless-fallback"],
     "codex-cli-bounded-visible-pilot-adapter": ["--first-response-fixture", "--idle-fixture"],
     "codex-cli-visible-first-response-capture-plan": ["--first-response-path", "--idle-path"],
     "codex-cli-visible-attach-acceptance": ["--proof-fixture", "--idle-fixture"],
+}
+VISIBLE_DRIVER_COMMANDS = {
     "codex-cli-visible-driver-plan": ["--fixture", "--codex-bin"],
     "codex-cli-local-driver-plan": ["--fixture", "--codex-bin"],
     "codex-cli-visible-driver-run": ["--proof-fixture", "--allow-headless-fallback"],
@@ -65,7 +69,9 @@ def require_json_success(result: subprocess.CompletedProcess[str]) -> dict[str, 
 
 def assert_source_shape() -> None:
     starter_source = STARTER.read_text(encoding="utf-8")
+    visible_common_source = VISIBLE_COMMON.read_text(encoding="utf-8")
     visible_driver_source = VISIBLE_DRIVER.read_text(encoding="utf-8")
+    visible_pilot_source = VISIBLE_PILOT.read_text(encoding="utf-8")
     runtime_idle_source = RUNTIME_IDLE.read_text(encoding="utf-8")
     init_source = INIT.read_text(encoding="utf-8")
 
@@ -96,6 +102,9 @@ def assert_source_shape() -> None:
     ):
         require(marker in starter_source, f"starter.py missing visible-driver delegation marker: {marker}")
 
+    for command in VISIBLE_PILOT_COMMANDS:
+        require(command in visible_pilot_source, f"starter_visible_pilot.py missing command: {command}")
+        require(command not in visible_driver_source, f"pilot command leaked into starter_visible_driver.py: {command}")
     for command in VISIBLE_DRIVER_COMMANDS:
         require(command in visible_driver_source, f"starter_visible_driver.py missing command: {command}")
 
@@ -106,6 +115,27 @@ def assert_source_shape() -> None:
         "_VISIBLE_DRIVER_HANDLERS",
     ):
         require(marker in visible_driver_source, f"starter_visible_driver.py missing marker: {marker}")
+    for marker in (
+        "register_starter_visible_pilot_commands(subparsers)",
+        "handle_starter_visible_pilot_command(args, print_payload)",
+    ):
+        require(marker in visible_driver_source, f"starter_visible_driver.py missing pilot delegation marker: {marker}")
+    for marker in (
+        "def register_starter_visible_pilot_commands(",
+        "def handle_starter_visible_pilot_command(",
+        "def handle_codex_cli_one_message_loop_pilot_command(",
+        "def handle_codex_cli_visible_attach_acceptance_command(",
+        "_VISIBLE_PILOT_HANDLERS",
+    ):
+        require(marker in visible_pilot_source, f"starter_visible_pilot.py missing marker: {marker}")
+
+    for marker in (
+        "def _add_project_arguments(",
+        "def _add_codex_probe_arguments(",
+        "def _add_optional_proof_fixture(",
+        "def _add_headless_fallback_argument(",
+    ):
+        require(marker in visible_common_source, f"starter_visible_common.py missing marker: {marker}")
 
     for marker in (
         "def _add_runtime_idle_observation_arguments(",
@@ -116,14 +146,19 @@ def assert_source_shape() -> None:
 
     for marker in (
         "handle_starter_visible_driver_command",
+        "handle_starter_visible_pilot_command",
         "register_starter_visible_driver_commands",
+        "register_starter_visible_pilot_commands",
         "handle_codex_cli_visible_driver_run_command",
     ):
         require(marker in init_source, f"__init__ omitted visible-driver export: {marker}")
 
+    require(len(visible_driver_source.splitlines()) <= 220, "starter_visible_driver.py exceeded size guard")
+    require(len(visible_pilot_source.splitlines()) <= 340, "starter_visible_pilot.py exceeded size guard")
+
 
 def assert_cli_surfaces() -> None:
-    for command, needles in VISIBLE_DRIVER_COMMANDS.items():
+    for command, needles in {**VISIBLE_PILOT_COMMANDS, **VISIBLE_DRIVER_COMMANDS}.items():
         help_text = require_success(run_cli(command, "--help"))
         for needle in needles:
             require(needle in help_text, f"{command} help omitted {needle}")
